more refactoring of environmemts code
