+++ src/python/pants/backend/javascript/subsystems/nodejs.py
@@ -13,9 +13,9 @@ from typing import ClassVar
 
 from nodesemver import min_satisfying
 
+from pants.core.environments.target_types import EnvironmentTarget
 from pants.core.util_rules import asdf, search_paths, system_binaries
 from pants.core.util_rules.asdf import AsdfPathString, AsdfToolPathsResult
-from pants.core.util_rules.environments import EnvironmentTarget
 from pants.core.util_rules.external_tool import (
     DownloadedExternalTool,
     ExternalToolRequest,
