+++ src/python/pants/core/util_rules/search_paths.py
@@ -10,7 +10,7 @@ from pathlib import Path
 from typing import Any, ClassVar
 
 from pants.base.build_environment import get_buildroot
-from pants.core.util_rules.environments import EnvironmentTarget
+from pants.core.environments.target_types import EnvironmentTarget
 from pants.engine.collection import DeduplicatedCollection
 from pants.engine.env_vars import EnvironmentVars
 from pants.engine.rules import Rule, _uncacheable_rule, collect_rules, rule
