+++ src/python/pants/engine/internals/platform_rules_test.py
@@ -9,15 +9,16 @@ from unittest.mock import Mock
 import pytest
 
 from pants.build_graph.address import Address
+from pants.core.environments.subsystems import EnvironmentsSubsystem
 from pants.core.environments.target_types import (
     DockerEnvironmentTarget,
     DockerImageField,
     DockerPlatformField,
+    EnvironmentTarget,
     LocalEnvironmentTarget,
     RemoteEnvironmentTarget,
     RemotePlatformField,
 )
-from pants.core.util_rules.environments import EnvironmentsSubsystem, EnvironmentTarget
 from pants.engine.env_vars import CompleteEnvironmentVars, EnvironmentVars, EnvironmentVarsRequest
 from pants.engine.environment import EnvironmentName
 from pants.engine.internals.platform_rules import (
